Rename own use to drug
