+++ core/src/main/java/com/rettichlp/unicacityaddon/base/config/UnicacityAddonConfiguration.java
@@ -7,7 +7,7 @@ import com.rettichlp.unicacityaddon.base.config.join.CommandConfiguration;
 import com.rettichlp.unicacityaddon.base.config.join.PasswordConfiguration;
 import com.rettichlp.unicacityaddon.base.config.message.MessageConfiguration;
 import com.rettichlp.unicacityaddon.base.config.nametag.NameTagConfiguration;
-import com.rettichlp.unicacityaddon.base.config.ownUse.OwnUseConfiguration;
+import com.rettichlp.unicacityaddon.base.config.drug.DrugConfiguration;
 import com.rettichlp.unicacityaddon.base.config.reinforcement.ReinforcementConfiguration;
 import com.rettichlp.unicacityaddon.base.config.sloc.SlocConfiguration;
 import com.rettichlp.unicacityaddon.base.config.tablist.TabListConfiguration;
@@ -35,7 +35,7 @@ public interface UnicacityAddonConfiguration {
 
     EquipConfiguration equip();
 
-    OwnUseConfiguration ownUse();
+    DrugConfiguration drug();
 
     // message
