+++ core/src/main/java/com/rettichlp/unicacityaddon/commands/faction/badfaction/OwnUseCommand.java
@@ -2,9 +2,9 @@ package com.rettichlp.unicacityaddon.commands.faction.badfaction;
 
 import com.rettichlp.unicacityaddon.UnicacityAddon;
 import com.rettichlp.unicacityaddon.base.builder.TabCompletionBuilder;
-import com.rettichlp.unicacityaddon.base.config.ownUse.cocaine.Cocaine;
-import com.rettichlp.unicacityaddon.base.config.ownUse.marijuana.Marijuana;
-import com.rettichlp.unicacityaddon.base.config.ownUse.methamphetamin.Methamphetamin;
+import com.rettichlp.unicacityaddon.base.config.drug.cocaine.Cocaine;
+import com.rettichlp.unicacityaddon.base.config.drug.marijuana.Marijuana;
+import com.rettichlp.unicacityaddon.base.config.drug.methamphetamin.Methamphetamin;
 import com.rettichlp.unicacityaddon.base.enums.faction.DrugPurity;
 import com.rettichlp.unicacityaddon.base.enums.faction.DrugType;
 import com.rettichlp.unicacityaddon.base.registry.UnicacityCommand;
@@ -29,9 +29,9 @@ public class OwnUseCommand extends UnicacityCommand {
 
     @Override
     public boolean execute(String[] arguments) {
-        Cocaine cocaine = this.unicacityAddon.configuration().ownUse().cocaine();
-        Marijuana marijuana = this.unicacityAddon.configuration().ownUse().marijuana();
-        Methamphetamin methamphetamin = this.unicacityAddon.configuration().ownUse().methamphetamin();
+        Cocaine cocaine = this.unicacityAddon.configuration().drug().cocaine();
+        Marijuana marijuana = this.unicacityAddon.configuration().drug().marijuana();
+        Methamphetamin methamphetamin = this.unicacityAddon.configuration().drug().methamphetamin();
 
         List<String> commandQueue = new ArrayList<>();
